cast to int

+++ ethereum/ethstats/src/main/java/org/hyperledger/besu/ethstats/EthStatsService.java
@@ -424,8 +424,9 @@ public class EthStatsService {
     final boolean isSyncing = syncState.isInSync();
     final long gasPrice = suggestGasPrice(blockchainQueries.getBlockchain().getChainHeadBlock());
     final long hashrate = miningCoordinator.hashesPerSecond().orElse(0L);
-    final long peersNumber =
-        protocolManager.ethContext().getEthPeers().streamAvailablePeers().count();
+    // safe to cast to int since it isn't realistic to have more than max int peers
+    final int peersNumber =
+        (int) protocolManager.ethContext().getEthPeers().streamAvailablePeers().count();
 
     final NodeStatsReport nodeStatsReport =
         ImmutableNodeStatsReport.builder()
